LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/auto-margins-position-area-expected.txt


FAIL .abspos 1 assert_equals:
<div class="container">
  <div class="anchor"></div>
  <div class="abspos" data-offset-x="10" data-offset-y="65">
    <div style="width: 80px; height: 20px;"></div>
  </div>
</div>
offsetLeft expected 10 but got 0
